+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Do not stop follow importer task on validation error.
+- Fixed PieFed webfinger address resolution.
 
 ## [2.18.0] - 2024-05-07